LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/inline-size-and-min-width-expected.txt

Green

FAIL min-width of inline-size container affects container size assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
